add git checkout for composer-sample-applications
Inserted steps 3 and 4 in Section 2 and bumped the subsequent step numbers appropriately.  Since the lab is hard-coded at version 0.11.3, I added the appropriate git checkout command to get the composer-sample-application repo at the appropriate commit level so that the package.json file in the digitalproperty-app will install the right version (0.11.3) locally

Also, fixed a typo in Section 4 step 2 so that the image shows up

And, added the hard-coded specification of @0.11.3. to the npm global install of generator-hyperledger-composer